Various cleanup and preparing to open source
