+++ pixi.toml
@@ -3,6 +3,9 @@ name = "action"
 version = "0.1.0"
 description = "Automated Camera Trapping Identification and Organization Network (ACTION)"
 repository = "https://github.com/humphrem/action"
+readme = "README.md"
+license = "Apache-2.0"
+license-file = "LICENSE"
 authors = [ "Morgan Humphrey", "David Humphrey"]
 channels = ["conda-forge"]
 platforms = ["win-64", "linux-64", "osx-64", "osx-arm64"]
