+++ scripts/aquatic-demo.sh
@@ -1,3 +1,3 @@
-#!/bin/zsh
+#!/bin/bash
 
 python3 action.py ./video/aquatic-demo.mov -c 0.45 -m 3.0 -s -b 1.0 -d -e aquatic
