+++ src/base_detector.py
